修正了电机pwm控制函数的bug，增加了机械臂动作调试时的打印信息，修改了四个轮子的转向与编码器、PWM的对应关系

+++ Core/Inc/ControlTask.h
@@ -8,8 +8,8 @@
 #define yhMotorEncoder 3
 #define yqMotorEncoder 2
 
-extern int32_t lockFlag;
-extern int32_t AngleAndPositionTIM;
+extern uint8_t lockFlag;
+extern uint8_t AngleAndPositionTIM;
 
 extern uint16_t Time1_ms;
 extern uint16_t Time2_ms;
